+++ packages/store/src/index.ts
@@ -1,4 +1,6 @@
 export { Storage } from './storage';
+export { applyManifestFilters, loadFilterPlugins } from './filter-pipeline';
+export type { Filters } from './filter-pipeline';
 export * from './lib/storage-utils';
 export * from './lib/versions-utils';
 export * from './lib/star-utils';